refactor(editor): split the forma planner into focused modules
multi-variant-auto-forma.ts was 947 lines mixing combinatorics, the
greedy per-variant packer, candidate gathering, step-diffing, and the
stage 1→2→3 orchestration. Extract the cohesive pieces into leaf modules:

- auto-forma-types.ts       — shared types (+ SharedCapacityInput, IterBudget)
- auto-forma-combinatorics.ts — lazy permutation/combination generators
- auto-forma-greedy.ts      — tryArrangeVariant + greedy normal-slot packer
- auto-forma-candidates.ts  — slot-spec / kind-bucket / forma-candidate gathering
- auto-forma-steps.ts       — diffFormaSteps + plan assembly

The entry (multi-variant-auto-forma.ts, now 385 lines) keeps only the
orchestration and re-exports the public surface, so existing imports
(editor-shell, auto-forma-dialog, tests) are unchanged. diffFormaSteps
now single-sources the before→after step diff that was duplicated in
stage 1 and the stage 2/3 plan builder.

Pure code-move: behavior identical, all 59 build-editor tests pass,
typecheck/lint/fmt clean. No import cycles (leaves never import the entry).
